Commented out Logic
As per Maria's comment, I am moving the logic we previously added to this file into the parse_gatorgrader_arguments.py file. I have temporally commented out this logic in this file, rather than deleting it. I will delete it from this file once it is cleared in the next file.