Proposed issueTemplate.md File for Issue #125
This is a manually created template for a single issue in the GatorGrouper repository. The code in the above file should be followed when raising an issue in the Issue Tracker to ensure readability for TA's and Professors. This will allow us to maintain a certain standard and raise issues in a similar manner collectively.