Adding License for the Project
In Issue #138, we decided to use the GNU General Public License. We selected this software license because it was free and open source.  This license guarantees users the freedom to run, study, share and modify the software. It has the permissions of Commercial Use, Distribution, Modification, Patent Use, and Private Use. Although the Limitations include Liability and Warranty.